+++ CHANGELOG.md
@@ -23,6 +23,8 @@ The version is based on the year, month, day, and release number. They are liste
 # v26.1.24
 * **Snap**: Resolved compatibility issues with snap applications;
 * **Manjaro/Arch**: Fixed package creation workflow;
+* **AppImage**: Added check for Node.js version;
+* **Debian**: Enforced dependency on Node.js >= 22.
 
 # v26.1.21
 * **Refactoring nest**: `iso` directory was moved inside `mnt/iso` to free more space for remastering, (look [here](https://github.com/pieroproietti/penguins-eggs/blob/master/conf/README.md));